Create PhoneBilling umbrella project
The PhoneBilling receives call detail records and calculates monthly bills for a given telephone number.

The PhoneBilling is structured as three focused applications:

BillingGateway - Is the web API requests processor. It was created as a Phoenix application.

BillingProcessor - Apply the business rules of telephone calls billing calculation.

BillingRepository - An application to persist and recover telephone calls records. It is using Ecto and Postgresql as the persistence strategy.

The PhoneBilling was separated in three applications in order to provide understandability, flexibility and manutenibility through the project lifecycle.