Get all duplicated call records from database
A call has a start record and end record. To identify a call
the start and end records needs to have the same call id. And to identify a call record it record need to have an unique id.

It will not be allowed two calls with the same call ids and two call records with the same id.

The purpose of this module is to search the existing identifiers and return when it is already inserted .