Unbind call records processed mounting steps response
The response is mounted by two lists, one that is produced by consistency validations and another that is the product of database insertions of valid calls.

The previous algorithm first do for each call record processed the correspondent response piece, only then the database insertions list would start its response mounting step.

With this change, it is possible to parallelize the steps, improving processing speed.